İsterler gerçekleştirildi.
Font eklendi.
Screenler düzenlendi.
Widgetler düzenlendi.
Theme dinamikleştirildi.
Bazı code refactoringler yapıldı.
