+++ lib/widgets/application_bar.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class ApplicationBar extends StatelessWidget implements PreferredSizeWidget {
+  const ApplicationBar({required this.appBar, super.key});
+
+  final AppBar appBar;
+
+  @override
+  Widget build(BuildContext context) {
+    return AppBar(
+      backgroundColor: Colors.transparent,
+    );
+  }
+
+  @override
+  // TODO: implement preferredSize
+  Size get preferredSize => Size.fromHeight(appBar.preferredSize.height);
+}
