+++ lib/widgets/drawer_menu_widget.dart
@@ -6,7 +6,6 @@ class SideMenu extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Drawer(
-
       child: ListView(
         padding: const EdgeInsets.all(8),
         children: [
@@ -14,8 +13,9 @@ class SideMenu extends StatelessWidget {
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
               SizedBox(
-                height: 70,
-                child: DrawerHeader(child: Image.network('https://tobeto.com/_next/image?url=%2F_next%2Fstatic%2Fmedia%2Ftobeto-logo.409772fc.png&w=256&q=75'),
+                height: 100,
+                child: DrawerHeader(
+                  child: Image.asset('assets/images/tobeto-logo.png'),
                 ),
               ),
             ],
@@ -46,9 +46,7 @@ class SideMenu extends StatelessWidget {
               trailing: Icon(Icons.person),
             ),
           ),
-          const ListTile(
-            title: Text('\u00a9 2023 Tobeto')
-          ),
+          const ListTile(title: Text('\u00a9 2023 Tobeto')),
         ],
       ),
     );
