+++ pubspec.yaml
@@ -74,11 +74,12 @@ flutter:
   # list giving the asset and other descriptors for the font. For
   # example:
   fonts:
-    - family: Tinos
-      fonts: 
-         - asset: assets/fonts/Tinos-Regular.ttf
-         - asset: assets/fonts/Tinos-Italic.ttf
-  #         style: italic
+      - family: Tinos
+        fonts:
+          - asset: assets/fonts/Tinos-Regular.ttf
+          - asset: assets/fonts/Tinos-Italic.ttf
+          - asset: assets/fonts/Tinos-Bold.ttf
+          - asset: assets/fonts/Tinos-BoldItalic.ttf
   #   - family: Trajan Pro
   #     fonts:
   #       - asset: fonts/TrajanPro.ttf
